+++ profiles/ulf/CHANGELOG.md
@@ -3,6 +3,7 @@
 # v2.5.x
 
 ## 2.5.4
+* SUPPORT-1040: Added Taxonomy term view for target group on courses
 * SUPPORT-1044: Changed contact form email text.
 
 ## 2.5.3